Lecture181 refactor index.ts
