+++ auth/src/app.ts
@@ -1,0 +1,35 @@
+import express from 'express'
+import 'express-async-errors'
+import { json } from 'body-parser'
+import cookieSession from 'cookie-session'
+
+import { currentUserRouter } from './routes/current-user'
+import { signinRouter } from './routes/siginin'
+import { signoutRouter } from './routes/signout'
+import { signupRouter } from './routes/signup'
+import { errorHandle } from './middlewares/error-handler'
+import { NotFoundError } from './errors/not-found-error'
+
+const app = express()
+app.set('trust proxy', true)
+app.use(json())
+
+app.use(
+	cookieSession({
+		signed: false,
+		secure: true,
+	})
+)
+
+app.use(currentUserRouter)
+app.use(signinRouter)
+app.use(signoutRouter)
+app.use(signupRouter)
+
+app.all('*', () => {
+	throw new NotFoundError()
+})
+
+app.use(errorHandle)
+
+export { app }
