+++ auth/src/index.ts
@@ -1,53 +1,23 @@
-import express from "express";
-import "express-async-errors"
-import { json } from "body-parser"
-import mongoose from "mongoose"
-import cookieSession from "cookie-session"
-
-import { currentUserRouter } from "./routes/current-user"
-import { signinRouter } from "./routes/siginin"
-import { signoutRouter } from "./routes/signout"
-import { signupRouter } from "./routes/signup"
-import { errorHandle } from "./middlewares/error-handler"
-import { NotFoundError } from "./errors/not-found-error"
-
-const app = express()
-app.set("trust proxy", true)
-app.use(json())
-
-app.use(cookieSession({
-    signed: false,
-    secure: true
-}))
-
-app.use(currentUserRouter)
-app.use(signinRouter)
-app.use(signoutRouter)
-app.use(signupRouter)
-
-app.all("*", () => {
-    throw new NotFoundError()
-})
-
-app.use(errorHandle)
+import mongoose from 'mongoose'
+import { app } from './app'
 
 const start = async () => {
-    if (!process.env.JWT_KEY) {
-        throw new Error("JWT_KEY must be defined")
-    }
+	if (!process.env.JWT_KEY) {
+		throw new Error('JWT_KEY must be defined')
+	}
 
-    try {
-        await mongoose.connect("mongodb://auth-mongo-srv:27017/auth", {
-            useNewUrlParser: true,
-            useUnifiedTopology: true,
-            useCreateIndex: true
-        })
-        console.log("Connected to mongodb")
-    } catch (err) {
-        console.error(err)
-    }
+	try {
+		await mongoose.connect('mongodb://auth-mongo-srv:27017/auth', {
+			useNewUrlParser: true,
+			useUnifiedTopology: true,
+			useCreateIndex: true,
+		})
+		console.log('Connected to mongodb')
+	} catch (err) {
+		console.error(err)
+	}
 
-    app.listen(3000, () => console.log("Listening on port 3000!!!"))
+	app.listen(3000, () => console.log('Listening on port 3000!!!'))
 }
 
 start()
